core/scraper.py
import httpx
import logging
from typing import List, Dict, Any
from bs4 import BeautifulSoup
import re

logger = logging.getLogger("artfish.pro.scraper")

class SkillDiscoveryScraper:
    """
    技能自动发现模块：通过爬虫自动寻找艺术类 API 并在系统中注册。
    """
    def __init__(self):
        self.target_sources = [
            "https://github.com/public-apis/public-apis",
            "https://www.programmableweb.com/category/art/apis"
        ]
        # 已识别的艺术类 API 清单 (示例)
        self.identified_apis = [
            {"name": "Art Institute of Chicago API", "url": "https://api.artic.edu/docs/", "type": "collection"},
            {"name": "Metropolitan Museum of Art API", "url": "https://metmuseum.github.io/", "type": "collection"},
            {"name": "Harvard Art Museums API", "url": "https://github.com/harvardartmuseums/api-docs", "type": "collection"},
            {"name": "DALL-E API", "url": "https://openai.com/blog/dall-e-api-now-available-in-public-beta", "type": "generation"},
            {"name": "MusicGraph API", "url": "https://developer.musicgraph.com/", "type": "music"}
        ]

    async def discover_new_skills(self) -> List[Dict[str, str]]:
        """
        自动爬取并发现新的艺术 API。
        """
        logger.info("Starting skill discovery crawl...")
        new_discoveries = []
        
        # 示例：模拟爬取过程
        try:
            # 在实际实现中，这里会解析 GitHub 或 API 聚合网站
            # async with httpx.AsyncClient() as client:
            #     for source in self.target_sources:
            #         resp = await client.get(source)
            #         # BeautifulSoup 解析逻辑...
            pass
        except Exception as e:
            logger.error(f"Discovery crawl failed: {e}")
            
        return self.identified_apis # 返回包含预置和新发现的清单

    def generate_api_manifest(self) -> str:
        """生成已识别的艺术类 API 清单文档"""
        manifest = "# Artfish Studio: 已识别的艺术类 API 清单\n\n"
        for api in self.identified_apis:
            manifest += f"- **{api['name']}**\n"
            manifest += f"  - URL: {api['url']}\n"
            manifest += f"  - 类型: {api['type']}\n\n"
        return manifest

# 实例
skill_scraper = SkillDiscoveryScraper()
