requirements.txt
# Core dependencies
typing-extensions>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API Layer
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
slowapi>=0.1.9
starlette>=0.27.0

# AI Models
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# Database Persistence
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Exporting results
reportlab>=4.0.0
svglib>=1.5.0

# Async Queue
celery>=5.3.0
redis>=5.0.0

# Distributed Communication
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# CLI & UX (Figma Grade Terminal)
rich>=13.0.0
typer[all]>=0.9.0
questionary>=2.0.0
psutil>=5.9.0
pyyaml>=6.0.0

# Claude MCP (Model Context Protocol)
mcp>=0.1.0

# Others
python-dotenv>=1.0.0
python-telegram-bot[job-queue]>=20.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
anyio>=4.0.0
bandit>=1.7.0
